requirements.txt
scipy
numpy
matplotlib
easydict
opencv-python
munkres
PyYAML
Shapely